Clients now show recently connected players properly.
This required some changes to how the game initiates the player-objects.
Instead of createing a new object every time a player joins, all objects will be initiated when a client start the game and then later customized when players join.
Another addition is that a client now observes changes to the room.players-array to be able add new players as soon as they join.