Improved client-helper updateSubscriptions
updateSubscriptions will now only re-subscribe to subscriptions
which are inactive. To determine which is inactive,
this helper now gets an list of the current subscriptions by looking
at Meteor.default_connection._subscriptions.

This fixes the issue that each client previously could open multiple
subscriptions of the same type, which wasted huge resources for both
 clients and server.